move asset selection code to utils file + also move createAndAppendElement
